fix: Rework Chat interface and fix incorrect Zod schema for structured outputs.
