+++ src/app.html
@@ -22,6 +22,11 @@
       type="image/png"
       href="%sveltekit.assets%/favicon-192.png"
     />
+    <link
+      rel="icon"
+      type="image/svg+xml"
+      href="%sveltekit.assets%/favicon.svg"
+    />
     <meta name="viewport" content="width=device-width" />
     <meta name="theme-color" content="#7E58FF" />
     <meta name="og:image" content="%sveltekit.assets%/og-image.png" />
