+++ src/lib/components/icons/ArrowUpIcon.svelte
@@ -1,0 +1,14 @@
+<svg
+  width="24"
+  height="24"
+  viewBox="0 0 24 24"
+  fill="none"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M12.7071 5.39644C12.3166 5.00592 11.6834 5.00592 11.2929 5.39644L6.21969 10.4697L5.68936 11L6.75002 12.0606L7.28035 11.5303L11.25 7.56065V18.25V19H12.75V18.25V7.56065L16.7197 11.5303L17.25 12.0606L18.3107 11L17.7803 10.4697L12.7071 5.39644Z"
+    fill="currentColor"
+  />
+</svg>
